+++ src/main/kotlin/org/rust/bsp/service/BspConnectionService.kt
@@ -19,7 +19,5 @@ interface BspConnectionService {
 
     fun disconnect()
 
-    fun doStuff()
-
     fun getProjectData(): CargoMetadata.Project
 }
